refactor(svelte2tsx): コンポーネント export 組み立てを add_component_export.rs へ
`$$render()` 本体の後ろに付く「async ラッパーの閉じ + `return { props, slots, events }`
+ component export」の組み立てを `addComponentExport.ts` 相当のファイルへ切り出す。

入力が多いので上流の `addComponentExport(params)` に倣って `ComponentExportParams`
構造体を用意し、関数先頭で分解している。本体のコードは 1 行も変えていない。

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
